shift stuff to serv_utils, to simplify code
